@@ -1,13 +1,24 @@
 package com.flight.management.config;
 
+import java.util.Arrays;
+import java.util.List;
+
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpMethod;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.web.cors.CorsConfiguration;
+import org.springframework.web.cors.CorsConfigurationSource;
+import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
 
 
 
+@SuppressWarnings("deprecation")
 @Configuration
 public class SpringSecurityConfig extends WebSecurityConfigurerAdapter {
 
@@ -15,13 +26,29 @@ public class SpringSecurityConfig extends WebSecurityConfigurerAdapter {
 	@Override
 	protected void configure(HttpSecurity http) throws Exception {
 		// TODO Auto-generated method stub
-		http.httpBasic().and().authorizeRequests().
-		antMatchers(HttpMethod.POST, "/**").hasAnyRole("ADMIN").and().csrf().disable().headers().
-		frameOptions().disable();
+		http.requestMatchers().
+		antMatchers(HttpMethod.OPTIONS, "/**").
+		and().csrf().disable().headers().
+		frameOptions().disable().and().cors().configurationSource(request-> {
+            CorsConfiguration configuration = new CorsConfiguration();
+            configuration.setAllowedOrigins(Arrays.asList("http://localhost:4200"));
+            configuration.setAllowedMethods(Arrays.asList("GET","POST"));
+            configuration.setAllowedHeaders(List.of("*"));
+            return configuration;
+        });
 	}
 	
-	@Override
-	protected void configure(AuthenticationManagerBuilder auth) throws Exception {
-		auth.inMemoryAuthentication().withUser("admin345").password("{noop}password").roles("ADMIN");
-	}
+	private CorsConfigurationSource configurationSource() {
+		  UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+		  CorsConfiguration config = new CorsConfiguration();
+		  config.addAllowedOrigin("*");
+		  config.setAllowCredentials(true);
+		  config.addAllowedHeader("X-Requested-With");
+		  config.addAllowedHeader("Content-Type");
+		  config.addAllowedMethod(HttpMethod.POST);
+		  source.registerCorsConfiguration("/logout", config);
+		  return source;
+		}
+	
+	
 }
