@@ -1,5 +1,7 @@
 package com.flight.management.controller;
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -8,6 +10,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.flight.management.entity.Airline;
@@ -18,7 +21,7 @@ import com.flight.management.service.AdminService;
 
 @RestController
 @RequestMapping("/api/v1.0/flight")
-@CrossOrigin(origins = "http://localhost:49238")
+@CrossOrigin(origins = "http://localhost:4200")
 public class AdminServiceController {
 
 	@Autowired
@@ -35,19 +38,6 @@ public class AdminServiceController {
 		System.out.println(newAirline + "airline added");
 		return newAirline;
 	}
-
-	@PostMapping("/admin/login")
-	public ResponseEntity<HttpStatus> adminLogin(@RequestBody User user) {
-		Boolean isValidAdmin = false;
-		if (user.getUsername().equals("admin123") && user.getPassword().equals("password")) {
-			isValidAdmin = true;
-		}
-		if (isValidAdmin) {
-			return ResponseEntity.ok(HttpStatus.ACCEPTED);
-		} else {
-			return ResponseEntity.ok(HttpStatus.BAD_GATEWAY);
-		}
-	}
 	
 	@PostMapping("/airline/inventory/add")
 	public ResponseEntity<String> addInventory(@RequestBody Flight flight) {
@@ -56,6 +46,14 @@ public class AdminServiceController {
 		System.out.println(newInventory+ "Flight added");
 		return ResponseEntity.ok(newInventory+" Flight added");
 
+	}
+	
+	@RequestMapping(value = "/getAllRegisteredAirline", method = RequestMethod.GET)
+	public List<Airline> getAllRegisteredAirline(){		
+		List<Airline> registeredAirlines = adminService.getAllRegisteredAirline();
+		System.out.println(registeredAirlines.size()+"size");
+		return registeredAirlines;
+		
 	}
 	
 
